refactor: methods out of DefaultMessageStore: GetMessageService
