+++ store/src/main/java/org/apache/rocketmq/store/plugin/AbstractPluginMessageStore.java
@@ -274,7 +274,7 @@ public abstract class AbstractPluginMessageStore implements MessageStore {
     @Override
     @Deprecated
     public boolean checkInDiskByConsumeOffset(String topic, int queueId, long consumeOffset) {
-        return next.checkInDiskByConsumeOffset(topic, queueId, consumeOffset);
+        return false;
     }
 
     @Override
